Basic custom page tables complete
Custom page table loading is now fully working in QEMU and on real
hardware.

With this came full ubsan support, and double buffering support to an
extent in the VGA code. It does require a pre-allocated buffer to be
parsed to the enabler function.

Many bug fixes also occurred. A major one being fixing assert so it is
consistent across C and C++ code within the kernel, whilst also
eliminating UB surrounding it.