Local APIC and i8259 PIC support
The legacy i8259 PICs are now remapped and essentially disabled.

There is now basic Local APIC initialisation support. With this came
some basic ACPI table parsing to discover information about the APICs.

Most of the architecture specific init code was moved to a Processor
class in order to make SMP start-up easier in the future. This meant
that the GDT is a member of this class instead of a global variable.

A missing copyright header was added to kernel/lib/include/spinlock.h