Beginnings of Scheduler + uACPI
This commit introduces a basic scheduler. It currently only fully
supports running on a single processor, but is designed to eventually
run on multiple. It is designed with user and kernel mode in mind, but
only supports kernel mode fully currently.

Some changes to the ISR routine were made as well to support processor
local data.

Support for uACPI was also added. It was originally added for MADT table
support, but it was discovered that it would be to complex at this
stage, so uACPI was left here with no actual use yet.